+++ packages/expo-dev-menu/ios/UITests/DevMenuTests.swift
@@ -26,6 +26,24 @@ class MockedExtension: DevMenuExtensionProtocol {
   }
 }
 
+class MockedDelegate: DevMenuDelegateProtocol {}
+
+class TestInterceptorFirstLaunch: DevMenuTestInterceptor {
+  @objc
+  var shouldShowAtLaunch: Bool { get { return true } }
+
+  @objc
+  var isOnboardingFinishedKey: Bool { get { return false } }
+}
+
+class TestInterceptorOnboardingFinished: DevMenuTestInterceptor {
+  @objc
+  var shouldShowAtLaunch: Bool { get { return false } }
+
+  @objc
+  var isOnboardingFinishedKey: Bool { get { return true } }
+}
+
 class DevMenuTests: XCTestCase {
   
   override func setUp() {
@@ -52,6 +70,40 @@ class DevMenuTests: XCTestCase {
     assertViewExists(text: "localhost:1234")
     assertViewExists(text: "JS Engine:")
     assertViewExists(text: "JavaScriptCore")
+  }
+
+  func test_dev_menu_auto_launch() {
+    DevMenuTestInterceptorManager.setTestInterceptor(TestInterceptorFirstLaunch())
+    DevMenuManager.configure(withBridge: UIMockedNOOPBridge(delegate: nil, launchOptions: nil))
+    waitForDevMenu()
+    DevMenuTestInterceptorManager.setTestInterceptor(nil);
+  }
+
+  func test_dev_menu_auto_launch_bypass() {
+    DevMenuTestInterceptorManager.setTestInterceptor(TestInterceptorFirstLaunch())
+    UserDefaults.standard.set(true, forKey: "EXDevMenuDisableAutoLaunch")
+    DevMenuManager.shared.readAutoLaunchDisabledState()
+
+    let label = UILabel()
+    label.text = "Test App"
+    label.bounds = CGRect(x: 0, y: 0, width: 100, height: 100)
+    label.accessibilityIdentifier = "TestApp"
+    UIApplication.shared.keyWindow!.rootViewController!.view.addSubview(label)
+    waitForView(tag: "TestApp")
+
+    DevMenuManager.configure(withBridge: UIMockedNOOPBridge(delegate: nil, launchOptions: nil))
+    runMainLoop(for: 1)
+
+    XCTAssertNil(DevMenuUIMatchers.findView(tag: DevMenuViews.mainScreen))
+    XCTAssertNil(DevMenuUIMatchers.findView(tag: DevMenuViews.footer))
+    XCTAssertFalse(DevMenuManager.shared.isVisible)
+
+    // clean up
+    DevMenuManager.shared.toggleMenu()
+    waitForDevMenu()
+    UserDefaults.standard.set(false, forKey: "EXDevMenuDisableAutoLaunch")
+    DevMenuManager.shared.readAutoLaunchDisabledState()
+    DevMenuTestInterceptorManager.setTestInterceptor(nil);
   }
   
   func test_if_dev_menu_can_be_toggled() {
@@ -96,11 +148,13 @@ class DevMenuTests: XCTestCase {
   }
   
   func test_if_menu_can_be_opened_on_settings_screen() {
+    DevMenuTestInterceptorManager.setTestInterceptor(TestInterceptorOnboardingFinished())
     DevMenuManager.configure(withBridge: UIMockedNOOPBridge(delegate: nil, launchOptions: nil))
     DevMenuManager.shared.openMenu("Settings")
 
     waitForView(tag: DevMenuViews.settingsScreen)
     waitForView(tag: DevMenuViews.footer)
     XCTAssertTrue(DevMenuManager.shared.isVisible)
+    DevMenuTestInterceptorManager.setTestInterceptor(nil);
   }
 }